+++ iblsorter/__init__.py
@@ -12,4 +12,4 @@ from .utils import Bunch, memmap_binary_file, read_data, load_probe, plot_dissim
 from .main import run, run_export, run_spikesort, run_preprocess
 from .io.probes import np1_probe, np2_probe, np2_4shank_probe
 
-__version__ = '1.9.1'
+__version__ = '1.10.0'