+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 Added / Removed / Changed / Fixed
 
-## [1.9.1 Unreleased]
+## [1.10.0] 2025-06-27
 
 ### Added
 - documentation and examples for EC2 and Docker integration